+++ Common.Mod.Generator/Specs/ConfigEntrySpec.cs
@@ -1,12 +1,15 @@
+using JetBrains.Annotations;
+
 namespace Common.Mod.Generator.Specs;
 
+[UsedImplicitly(ImplicitUseKindFlags.Assign, ImplicitUseTargetFlags.Members)]
 public record ConfigEntrySpec
 {
     public string Name { get; set; } = null!;
     public ConfigEntryTypeSpec Type { get; set; }
-    public object? DefaultValue { get; set; } = null;
-    public string? Enum { get; set; } = null;
-    public string? Nested { get; set; } = null;
-    public string? Label { get; set; } = null;
-    public string? Description { get; set; } = null;
+    public object? DefaultValue { get; set; }
+    public string? Enum { get; set; }
+    public string? Nested { get; set; }
+    public string? Label { get; set; }
+    public string? Description { get; set; }
 }