+++ Common.Mod.Generator/Specs/RootConfigSpec.cs
@@ -1,5 +1,8 @@
+using JetBrains.Annotations;
+
 namespace Common.Mod.Generator.Specs;
 
+[UsedImplicitly(ImplicitUseKindFlags.Assign, ImplicitUseTargetFlags.Members)]
 public record RootConfigSpec : ConfigSpec
 {
     public string Version { get; set; } = null!;