+++ Common.Mod.Generator/Specs/RootConfigTypeSpec.cs
@@ -1,5 +1,8 @@
+using JetBrains.Annotations;
+
 namespace Common.Mod.Generator.Specs;
 
+[UsedImplicitly(ImplicitUseKindFlags.Access, ImplicitUseTargetFlags.Members)]
 public enum RootConfigTypeSpec
 {
     Common,